PEP 570: Restructure sections as recommended by PEP 1
Organize PEP 570 according to the recommended ordering of sections
defined in PEP 1.  This presents the proposal in a more organized
fashion.

Specifically, the proposal reads in a more organized top-down fashion,
an explicit mention of API authors and callers are distinguished,
"parameters" and "arguments" are used correctly, and other grammatical
corrections and slight rewording.

While the diff looks drastic, the core content of the originally written
PEP remains in tact.